requirements.txt
matplotlib==3.10.3
moondream==0.1.0
Pillow==11.2.1
python-dotenv==1.1.0
Requests==2.32.4
